pipeline/requirements.txt
requests==2.32.3
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.2
pandas==2.2.2
numpy==2.0.1